app/requirements.txt
flask
opencv-python-headless
torch
ultralytics
numpy
flask-socketio
eventlet
requests>=2.31.0
python-dotenv
